+++ tests/test_as_people.py
@@ -11,6 +11,6 @@ class TestPeople(TestCaseWithLog):
         users = [u async for u in self.async_api.people.list_gen()]
         # try to provoke 429
         detail_tasks = [self.async_api.people.details(person_id=user.person_id)
-                        for user in users*10]
+                        for user in users*20]
         details = await asyncio.gather(*detail_tasks)
         print(f'got details for {len(details)} users')